Version 2.0.0 rewrite
This is a rather big rewrite that I worked on (and didnt commit in stages) over quite a long period.
There are changes to how the files are being handled (mainly through the splitting into several classes that handle file writing/reading and data/repository classes).
Also functional changes to how both twitch and youtube handle their respective tasks.
Twitch:
- now actually updates stream announcements for secondary streams as well

YouTube:
- changed video check logic to use the channels upload playlist instead of the channel itself
- now also handles livestreams properly
- now allows adding a new notification via the channels name by getting the channels playlist id automatically

The discord bot slash commands also had some changes in the form of a few new commands for editing/removing of notifications and similar.
The slash commands are now registered via the Guild.updateCommands endpoint.
Also the general file structure (and formatting) of the project has been changed (this kinda breaks the changelist but this is a necessary evil for a better organized future).